PJ/Bitcoin_Auto_Trade/requirements.txt
python-dotenv
openai
pyupbit
ta
selenium
webdriver-manager
Pillow